Move shutdown logic from MainMenuView to controller
